Cleaned up code, look at detailed comments.
a. Added a map showing Evanston on the Front Page. We shall later
change it to the user's current location

b. In index.html, added data-ajax = false to the page link (for map and
havewi-fi) to enable reloads. For now takes care of our issue of having
to press refresh.

c. In havewi-fi, made a bunch of changes. Removed location input and
fetching current location instead, changed the button type for submit
to the one used by jquery to make the page work more reliably. Finally
just fiddled with the theme to make it match the rest of our pages.