Showing mail information on map and other changes
1. E-mail link shows up in the maps (maps.html) page from the Users
table.

2. Added changes to havewifi.html to to the following:
    - Slider for selecting current location or enter address.
    - If current location is selected, reverse geocoding is done and
the address is stored in Parse. The user is prompted to share location
if and only if he chooses to.
    - If current location is not chosen, the address is geocoded using
the current location (already implemented by Nick, Tarun and Yu, just
restructured the code slightly to reduce code duplication when handling
the current location feature listed above.

  - For map.html and havewifi.html, added Google Maps api for finding
address from I.P. address. For index.html, this reduces the need to ask
for current location, since the map is just supposed to give an idea of
location and not supposed to be exact.

- Added a button for a deleting wi-fi on the main screen. This is a
placeholder, it doesn't do anything. We should think about how to
remove listings, we can use the Wi-Fi name and e-mail to confirm
deletions.